+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Updated
 * Pipeline model supports null catalog (hive metastore)
 * Event Data Source supports custom event root path
+* Event Data Source supports custom schema location path
 
 ## [0.1.1] - 2024-01-28
 ### Added